Added Dynamic Folder and some more problems in it
