+++ Dynamic Programming/Binary Tree Right Side View.cpp
@@ -1,0 +1,37 @@
+/**
+ * Definition for a binary tree node.
+ * struct TreeNode {
+ *     int val;
+ *     TreeNode *left;
+ *     TreeNode *right;
+ *     TreeNode() : val(0), left(nullptr), right(nullptr) {}
+ *     TreeNode(int x) : val(x), left(nullptr), right(nullptr) {}
+ *     TreeNode(int x, TreeNode *left, TreeNode *right) : val(x), left(left), right(right) {}
+ * };
+ */
+class Solution {
+public:
+    vector<int> rightSideView(TreeNode* root) {
+        vector<int>ans;
+        if(!root){
+            return ans;
+        }
+        queue<TreeNode*>q;
+        
+        q.push(root);
+        
+        while(!q.empty()){
+            int n = q.size();
+            TreeNode* tmp;
+            for(int i = 0; i < n; i++){
+                tmp = q.front();
+                q.pop();
+                if(tmp->left) q.push(tmp->left);
+                if(tmp->right) q.push(tmp->right);
+            }
+            
+            ans.push_back(tmp->val);
+        }
+        return ans;
+    }
+};
