--- Pascal's Triangle.cpp
@@ -1,1 +1,0 @@
-
