--- README.md
@@ -1,2 +1,0 @@
-# This repo is for Hacktoberfest 2022.
-# DSA
